Add ability to show URN of selected text
The browser APIs for selection state are very finicky. See
https://developer.mozilla.org/en-US/docs/Web/API/Selection

But we're sort of working around this by using mouseup and mousedown
events to set the start and end nodes, and then using their
URNs to determine the correct direction for the citation.

It works reasonably well, but the mousedown needs to happen on a textnode.

We can try to find a way around this at some point, but it's likely
to take up a lot of time.